fixing workspace changes
