+++ cli/src/to/ascii.rs
@@ -6,9 +6,9 @@
  */
 
 use crate::GlobalArgs;
-use crate::graphs::bvgraph::get_endianness;
-use crate::traits::SequentialLabeling;
-use crate::utils::MmapHelper;
+use webgraph::graphs::bvgraph::get_endianness;
+use webgraph::traits::SequentialLabeling;
+use webgraph::utils::MmapHelper;
 use anyhow::Result;
 use clap::Parser;
 use dsi_bitstream::dispatch::factory::CodesReaderFactoryHelper;
@@ -44,7 +44,7 @@ pub fn ascii_convert<E: Endianness + 'static>(global_args: GlobalArgs, args: Cli
 where
     MmapHelper<u32>: CodesReaderFactoryHelper<E>,
 {
-    let seq_graph = crate::graphs::bvgraph::sequential::BvGraphSeq::with_basename(args.src)
+    let seq_graph = webgraph::graphs::bvgraph::sequential::BvGraphSeq::with_basename(args.src)
         .endianness::<E>()
         .load()?;
