+++ webgraph/Cargo.toml
@@ -18,6 +18,8 @@ fuzz = [
 	"dsi-bitstream/fuzz",
 ] # Expose the fuzzing harnesses
 serde = ["dep:serde"]
+be_bins = [] # Enable read / write of only BE bvgraphs (to reduce code size)
+le_bins = [] # Enable read / write of only LE bvgraphs (to reduce code size)
 
 [dependencies]
 anyhow.workspace = true
@@ -34,9 +36,9 @@ itertools.workspace = true
 rayon.workspace = true
 mmap-rs.workspace = true
 predicates.workspace = true
+java-properties.workspace = true
 arbitrary = { version = "1.3.2", features = ["derive"], optional = true }
 
-java-properties = "2.0.0"
 num_cpus = "1.16.0"
 libc = "0.2.155"
 impl-tools = "0.10.0"
@@ -49,7 +51,6 @@ crossbeam-channel = "0.5"
 
 # Fuzzing deps
 zip = { version = "2.1.5", optional = true }
-sysinfo = "0.33.1"
 sync-cell-slice = "0.9.9"
 
 [dev-dependencies]
